config: move ChangePasswordRequest into request folder
