+++ src/main/java/com/group3979/badmintonbookingbe/api/ProfileAPI.java
@@ -1,6 +1,6 @@
 package com.group3979.badmintonbookingbe.api;
 
-import com.group3979.badmintonbookingbe.model.ChangePasswordRequest;
+import com.group3979.badmintonbookingbe.model.request.ChangePasswordRequest;
 import com.group3979.badmintonbookingbe.model.request.ProfileRequest;
 import com.group3979.badmintonbookingbe.model.response.ProfileResponse;
 import com.group3979.badmintonbookingbe.service.ProfileService;
