+++ src/main/java/com/group3979/badmintonbookingbe/model/request/ChangePasswordRequest.java
@@ -1,4 +1,4 @@
-package com.group3979.badmintonbookingbe.model;
+package com.group3979.badmintonbookingbe.model.request;
 
 import lombok.Data;
